+++ dataql/solvers/registry.py
@@ -21,6 +21,7 @@ Example
 """
 
 from collections import Mapping
+from inspect import isfunction, ismethod
 
 from dataql.solvers.base import AttributeSolver, ObjectSolver, ListSolver
 from dataql.solvers.exceptions import CannotSolve
@@ -58,6 +59,14 @@ class Attribute:
         If ``args`` or ``kwargs`` are passed, the attribute got from ``value`` must be callable.
         They will be passed anyway (using ``[]`` and ``{}`` as default values) if the attribute
         is a function.
+
+        And if it's not a function but a real attribute, if no args are passed but the attribute
+        got from ``value`` is "callable", we'll try to call it without argument.
+        If any exception occurs during the call, two things may happen:
+        - the "thing" called is a function or method, then we raise the exception
+        - it's something else, like an instance of a class having a ``__call__`` method,
+          in this case we want to return the instance (if the user want the thing to be
+          called it could still use parentheses)
 
         Arguments
         ---------
@@ -101,7 +110,115 @@ class Attribute:
         >>> Attribute('str', function=str).solve(d)
         '2015-06-01'
 
+        Example with callable
+        ----------------------
+
+        >>> class Klass1:
+        ...     def __call__(self):
+        ...         return 'foo'
+        >>> class Klass2:
+        ...     def __call__(self, arg):
+        ...         return 'foo'
+        >>> class Klass3:
+        ...     def __call__(self):
+        ...         raise TypeError('error')
+        >>> class Klass4:
+        ...     def __call__(self, **kwargs):
+        ...         return 'foo'
+        >>> class Klass5:
+        ...     def __call__(self, **kwargs):
+        ...         return kwargs['foo']
+        >>> class Klass6:
+        ...     def __call__(self, *args):
+        ...         return 'foo'
+        >>> class Klass7:
+        ...     def __call__(self, *args):
+        ...         return args[0]
+        >>> class Klass:
+        ...     def call1(self):
+        ...         return 'foo'
+        ...     def call2(self, arg):
+        ...         return 'foo'
+        ...     def call3(self):
+        ...         raise TypeError('error')
+        ...     def call4(self, **kwargs):
+        ...         return 'foo'
+        ...     def call5(self, **kwargs):
+        ...         return kwargs['foo']
+        ...     def call6(self, *args):
+        ...         return 'foo'
+        ...     def call7(self, *args):
+        ...         return args[0]
+        ...     k1, k2, k3, k4 = Klass1(), Klass2(), Klass3(), Klass4()
+        ...     k5, k6, k7 = Klass5(), Klass6(), Klass7()
+
+        >>> v = Klass()
+
+        # Can be called without arguments: we call it
+        >>> Attribute('call1').solve(v)
+        'foo'
+
+        # TypeError while calling without argument, but it's a method: raise
+        >>> Attribute('call2').solve(v)
+        Traceback (most recent call last):
+        TypeError: call2() missing 1 required positional argument: 'arg'
+
+        # TypeError but no argument needed, but it's a method: raise
+        >>> Attribute('call3').solve(v)
+        Traceback (most recent call last):
+        TypeError: error
+
+        # Can be called without used kwargs: we call it
+        >>> Attribute('call4').solve(v)
+        'foo'
+
+        # KeyError while calling without argument, in a get of kwargs, but it's a method: raise
+        >>> Attribute('call5').solve(v)
+        Traceback (most recent call last):
+        KeyError: 'foo'
+
+        # Can be called without used args: we call it
+        >>> Attribute('call6').solve(v)
+        'foo'
+
+        # IndexError while calling without argument, in a get of args, but it's a method: raise
+        >>> Attribute('call7').solve(v)
+        Traceback (most recent call last):
+        IndexError: tuple index out of range
+
+
+        # Can be called without arguments: we call it
+        >>> Attribute('k1').solve(v)
+        'foo'
+
+        # TypeError while calling without argument, but it's NOT a method: return not called
+        >>> isinstance(Attribute('k2').solve(v), Klass2)
+        True
+
+        # TypeError but no argument needed, but it's NOT a method: return not called
+        >>> isinstance(Attribute('k3').solve(v), Klass3)
+        True
+
+        # Can be called without used kwargs: we call it
+        >>> Attribute('k4').solve(v)
+        'foo'
+
+        # KeyError while calling without argument, in a get of kwargs, but it's NOT a method:
+        # return not called
+        >>> isinstance(Attribute('k5').solve(v), Klass5)
+        True
+
+        # Can be called without used args: we call it
+        >>> Attribute('k6').solve(v)
+        'foo'
+
+        # IndexError while calling without argument, in a get of args, but it's NOT a method:
+        # return not called
+        >>> isinstance(Attribute('k7').solve(v), Klass7)
+        True
+
         """
+
 
         # If we have a function, apply this, using the value as first argument,
         # then args and kwargs.
@@ -111,13 +228,25 @@ class Attribute:
         # Manage a normal attribute.
         result = getattr(value, self.name)
 
+
         # We make a call from the attribute in all cases if we have some arguments.
-        # If we don't have arguments, we still check that the attribute is callable.
-        # If yes and if it's not a class (it will always answer yes for a class), then
-        # we call it.
-        has_args = args is not None or kwargs is not None
-        if has_args or (not isclass(result) and callable(result)):
+        if args is not None or kwargs is not None:
             result = result(*(args or []), **(kwargs or {}))
+
+        elif not isclass(result) and callable(result):
+            # We have no arguments, but the attribute is a callable, so, if it not a class,
+            # we'll try to call it, without arguments.
+            try:
+                result = result()
+
+            except Exception:
+                # If an exception is raised during the call, we have two cases:
+                # - the "thing" called is a function or method, then we re-raise the exception
+                # - it's something else, like an instance of a class having a ``__call__`` method,
+                #   in this case we want to return the instance (if the user want the thing to be
+                #   called it could still use parentheses)
+                if isfunction(result) or ismethod(result):
+                    raise
 
         return result